backend/pyproject.toml
[project]
name = "telegram-bot-backend"
version = "0.1.0"
description = "Backend: בוט טלגרם + Web App לחדר בריחה"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.6",
    "fastapi>=0.115.0",
    "pydantic>=2.0",
    "uvicorn[standard]>=0.32.0",
    "typing-extensions>=4.0",
    "google-genai>=1.0.0",
    "huggingface-hub>=0.28.0",
    "Pillow>=10.0.0",
    "requests>=2.28.0",
    "boto3>=1.28.0",
    "redis>=5.0.0",
    "sqlalchemy>=2.0.0",
    "pymysql>=1.1.0",
    "psycopg2-binary>=2.9.0",
]

[dependency-groups]
dev = []
